Move apiCalls to util folder, and set up reducer and action for setting movies to store
