+++ src/actions/index.js
@@ -1,0 +1,4 @@
+export const setMovies = movies => ({
+  type: 'SET_MOVIES',
+  movies
+})
